fix: address code review findings and improve README structure
- fix int/ssize_t mismatch in tun_read_packet and handlers
  - fix TCP partial write loop in net_tcp_send_frame
  - fix TCP EOF detection with stale errno
  - fix reconnect replay rejection by resetting recv_seq_max
  - fix missing server_reset_replay and client_id in manual reconnect
  - add sys/types.h include for ssize_t
  - update README: badges, Quick Start, Features, improved layout

+++ camex.c
@@ -382,7 +382,7 @@ static void drain_udp_packets(void)
 {
     uint8_t buffer[TUN_PACKET_MAX + CAMEX_HDR_LEN + 1 + 16U];
     struct sockaddr_in from;
-    int len;
+    ssize_t len;
 
     if (server_mode) {
         while (1) {
@@ -538,7 +538,7 @@ static void drain_tcp_client_read(int fd)
 static void handle_tun_packet(void)
 {
     uint8_t buffer[TUN_PACKET_MAX];
-    int len;
+    ssize_t len;
     uint32_t src_ip_be = 0;
     uint32_t dst_ip_be = 0;
